--- {{cookiecutter.project_slug}}/server/{{cookiecutter.project_slug}}/core/handlers.py
@@ -1,9 +1,0 @@
-from django.conf import settings
-from storages.backends.s3boto3 import S3Boto3Storage
-
-
-class PrivateMediaStorage(S3Boto3Storage):
-    location = settings.PRIVATE_MEDIAFILES_LOCATION
-    default_acl = "private"
-    file_overwrite = False
-    custom_domain = False
